backend/requirements.txt
requests
beautifulsoup4
selenium
pytesseract
Pillow
openai 
fastapi
uvicorn

